added damage factors for bullets
git-svn-id: file:///C/Users/nicol/Documents/GitHub/SVN/GGJ19@186 64d83045-0e65-4e92-91a9-31b6b1dfce6a

+++ Project/Assets/Scripts/Systems.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Systems : MonoBehaviour
 {
-    public float WeaponLvL = 1;
+    public int WeaponLvL = 1;
     public int Health = 100;
     public float DriveFactor = 1;
